+++ src/main/java/com/lvl6/server/controller/RedeemUserCityGemsController.java
@@ -205,7 +205,7 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
       Map<Integer, Integer> boosterPackIdsToQuantities = 
           new HashMap<Integer, Integer>();
       boosterPackIdsToQuantities.put(boosterPackId, expectedNumEquips);
-      List<Integer> userEquipIds = new ArrayList<Integer>();
+      List<Long> userEquipIds = new ArrayList<Long>();
       
       //WARNING: this will contain booster items from multiple booster
       //packs if redeeming gems will be for more than one booster pack
@@ -242,7 +242,7 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
   
   //return list of full user equip protos and populate userEquipIds
   private List<FullUserEquipProto> purchaseBoosterPacks(User u,
-      Map<Integer, Integer> boosterPackIdsToQuantities, List<Integer> userEquipIds,
+      Map<Integer, Integer> boosterPackIdsToQuantities, List<Long> userEquipIds,
       Map<Integer, Integer> sendToClientBoosterItemIdsToQuantities) {
     
     List<FullUserEquipProto> returnValue = new ArrayList<FullUserEquipProto>();
@@ -343,7 +343,7 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
   private void recordPurchases(int userId, Timestamp now,
       Map<Integer, List<BoosterItem>> packIdToItemsUserReceives,
       Map<Integer, Integer> boosterPackIdsToQuantities,
-      List<Integer> userEquipIds) {
+      List<Long> userEquipIds) {
     //this one won't count towards the daily limit
     boolean excludeFromLimitCheck = true;
 
@@ -363,10 +363,10 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
       Map<Integer, Integer> boosterItemIdsToNumCollected,
       Map<Integer, Integer> newBoosterItemIdsToNumCollected,
       List<BoosterItem> itemsUserReceives, List<Boolean> collectedBeforeReset,
-      boolean resetOccurred, List<Integer> newUserEquipIds, Timestamp now) {
+      boolean resetOccurred, List<Long> newUserEquipIds, Timestamp now) {
     int userId = aUser.getId();
     String reason = ControllerConstants.UER__REDEEM_USER_CITY_GEMS;
-    List<Integer> userEquipIds = MiscMethods.insertNewUserEquips(userId,
+    List<Long> userEquipIds = MiscMethods.insertNewUserEquips(userId,
         itemsUserReceives, now, reason);
     if (null == userEquipIds || userEquipIds.isEmpty() || userEquipIds.size() != itemsUserReceives.size()) {
       log.error("unexpected error: failed to insert equip for user. boosteritems="
@@ -388,14 +388,14 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
   }
   
   private List<FullUserEquipProto> constructFullUserEquipProtos(int userId, 
-      List<BoosterItem> biList, List<Integer> userEquipIds) {
+      List<BoosterItem> biList, List<Long> userEquipIds) {
     List<FullUserEquipProto> returnValue = new ArrayList<FullUserEquipProto>();
     
     int forgeLevel = ControllerConstants.DEFAULT_USER_EQUIP_LEVEL;
     int enhancementLevel = ControllerConstants.DEFAULT_USER_EQUIP_ENHANCEMENT_PERCENT;
     for (int i = 0; i < biList.size(); i++) {
       BoosterItem bi = biList.get(i);
-      int userEquipId = userEquipIds.get(i);
+      long userEquipId = userEquipIds.get(i);
       int equipId = bi.getEquipId();
       int durability = ControllerConstants.DEFAULT_USER_EQUIP_DURABILITY;
